requirements.txt
google-genai
groq
youtube-transcript-api
python-dotenv
rich
click
streamlit
reportlab